# Conflicts:
#	android/app/src/main/kotlin/com/gemwallet/android/PaymentNavigation.kt
#	android/app/src/test/kotlin/com/gemwallet/android/PaymentNavigationTest.kt
#	android/features/confirm/viewmodels/src/main/kotlin/com/gemwallet/android/features/confirm/viewmodels/ConfirmViewModel.kt
#	android/features/confirm/viewmodels/src/main/kotlin/com/gemwallet/android/features/confirm/viewmodels/models/ConfirmHeaderUIModel.kt
#	android/gemcore/src/main/kotlin/com/gemwallet/android/domains/confirm/AmountUIModel.kt
#	android/gemcore/src/test/kotlin/com/gemwallet/android/domains/confirm/AmountUIModelTest.kt
#	android/gemcore/src/testFixtures/kotlin/com/gemwallet/android/testkit/AmountUIModelMock.kt
#	android/ui/src/main/kotlin/com/gemwallet/android/ui/components/list_item/GemListRowView.kt
#	android/ui/src/main/kotlin/com/gemwallet/android/ui/components/list_item/ListItemModel.kt
#	core/gemstone/src/payment.rs
#	core/gemstone/src/services/recipient/rules.rs
#	ios/Features/Transfer/Sources/Types/PaymentDestinationBuilder.swift
#	ios/Features/Transfer/Sources/ViewModels/ConfirmHeaderViewModel.swift
#	ios/Features/Transfer/Sources/ViewModels/ConfirmTransferSceneViewModel.swift
#	ios/Features/Transfer/Sources/ViewModels/TransactionInputViewModel.swift
#	ios/Features/Transfer/Tests/ViewModels/ConfirmHeaderViewModelTests.swift
#	ios/Gem/Navigation/NavigationRouter.swift
#	ios/Packages/GemstonePrimitives/TestKit/GemTransferServiceMocks.swift
#	ios/Packages/PrimitivesComponents/Sources/Extensions/Gemstone+Localized.swift
#	ios/Packages/PrimitivesComponents/Sources/Types/TransactionHeaderBuilder.swift
#	ios/Packages/PrimitivesComponents/Sources/ViewModels/TransactionInfoViewModel.swift

+++ android/features/transfer_amount/viewmodels/src/main/kotlin/com/gemwallet/android/features/transfer_amount/viewmodels/AmountViewModel.kt
@@ -179,7 +179,7 @@ class AmountViewModel @Inject constructor(service: GemAmountServiceInterface, fa
         stakeProvider?.setResource(resource)
     }
 
-    fun selectLeverage(value: Int) {
+    fun selectLeverage(value: UByte) {
         perpetualProvider?.setLeverage(value)
     }
